Add instructions flags; Now each instruction changes the internal status flag.

+++ include/emu/operand.h
@@ -23,6 +23,6 @@ typedef struct {
 	};
 } Operand;
 
-u32 operandGetData(CPU *cpu, Operand *op);
+u16 operandGetData(CPU *cpu, Operand *op);
 
 #endif /* _EMU_OPERAND_H_ */
